Crate dkg_runtime

Modules

api

xcm_config

Macros

construct_runtime

Construct a runtime, with the given name and the given pallets.

match_type

Create a type which implements the Contains trait for a particular type with syntax similar to matches!.

parameter_types

Create new implementations of the Get trait.

Structs

BlockExecutionWeight

Importing a block with 0 txs takes ~5 ms

BlockHashCount

ChainIdentifier

CreationFee

DKGAccountId

EnsureOneOf

The “OR gate” implementation of EnsureOrigin.

ExecutiveBody

ExistentialDeposit

ExtrinsicBaseWeight

Executing 10,000 System remarks (no-op) txs takes ~1.26 seconds -> ~125 µs per tx

GenesisConfig

IdentityFee

Implementor of WeightToFeePolynomial that maps one unit of weight to one unit of fee.

IsInVec

Trivial utility for implementing Contains/OrderedMembership with a Vec.

MaxAuthorities

MaxCandidates

MaxInvulnerables

MaxLocks

MaxReserves

MinCandidates

MinimumPeriod

Offset

OperationalFeeMultiplier

Origin

The runtime origin type representing the origin of a call.

PalletId

A pallet identifier. These are per pallet and should be stored in a registry somewhere.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill

A fixed point representation of a number in the range [0, 1].

Percent

A fixed point representation of a number in the range [0, 1].

Period

Permill

A fixed point representation of a number in the range [0, 1].

PotId

ProposalLifetime

RefreshDelay

ReservedDmpWeight

ReservedXcmpWeight

RocksDbWeight

By default, Substrate uses RocksDB, so this will be the weight used throughout the runtime.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

SS58Prefix

SessionKeys

SessionLength

TimeToRestart

TransactionByteFee

TransferFee

UncleGenerations

Version

WeightToFee

Handles converting a weight scalar to a fee value, based on the scale and granularity of the node’s balance type.

Enums

BalancesCall

Contains one variant per dispatchable that can be called by an extrinsic.

Call

DispatchClass

A generalized group of dispatch types.

Event

Everything

A [Contains] implementation that contains every value.

MultiAddress

A multi-format address wrapper for on-chain accounts.

OriginCaller

TimestampCall

Contains one variant per dispatchable that can be called by an extrinsic.

Constants

DAYS

EPOCH_DURATION_IN_BLOCKS

EXISTENTIAL_DEPOSIT

HOURS

MICROUNIT

MILLISECS_PER_BLOCK

This determines the average expected block time that we are targeting. Blocks will be produced at a minimum duration defined by SLOT_DURATION. SLOT_DURATION is picked up by pallet_timestamp which is in turn picked up by pallet_aura to implement fn slot_duration().

MILLIUNIT

MINUTES

SLOT_DURATION

UNIT

VERSION

This runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

WEIGHT_PER_SECOND

Traits

BuildStorage

Complex storage builder stuff.

Randomness

A trait that is able to provide randomness.

StorageValue

A trait for working with macro-generated storage values under the substrate storage API.

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirst

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Aura

AuraConfig

AuraExt

AuraExtConfig

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Authorship

Balance

Balance of an account.

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

CheckedExtrinsic

Extrinsic type that has already been checked.

CollatorSelection

CollatorSelectionConfig

CollatorSelectionUpdateOrigin

CumulusXcm

DKG

DKGConfig

DKGId

Identity of a DKG authority using ECDSA as its crypto.

DKGProposalHandler

DKGProposals

DKGProposalsConfig

DmpQueue

Executive

Executive: handles dispatch to the various modules.

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

Index

Index of a transaction in the chain.

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

RandomnessCollectiveFlip

Session

SessionConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Weight

Numeric range of a transaction weight.

XcmpQueue